requirements.txt
yt-dlp>=2023.3.4
Pillow>=9.0.0
requests>=2.28.0
qrcode>=7.3.1 